Switch Sparkfun SAMD51 MicroMod to internal flash
The Sparkfun SAMD51 MicroMod module ships with a 16MB external flash
chip, but the v1.2 version of the board has a design flaw which
prevents the SPI to the chip from working (PA10 cannot be MOSI).

The switches to the SAMD51's smaller internal flash, and reduces the
CircuitPython build to the minimum, similar to other boards that lack
external flash.