espressif: Add SDIOIO support
This still needs cleanup, but it does build and work with on a custom board.
That board only runs 1-bit SDIO mode.

I will try this out with a Memento board I have, but it won't really be useful there:
you lose the SPI bus.

Next step: find an officially supported board that has the right lines
for SDMMC support and/or a SD breakout board.