set errno for USBError & USBTimeoutError
This sets unique errno values for each place in the usb.core.Device
implementation that raises USBError or USBTimeoutError. The idea is
to provide more visibility so CircuitPython code can understand
what's gone wrong when TinyUSB complains about something.

CAUTION: This removes two error strings (stalled pipe and config
not set).

The new way relies on unique error codes as defined in
shared-module/usb/core/Device.h. The "errno" values here are
arbitrary with no relation to POSIX errno codes. I don't see this
as a problem because the so-called errno codes returned by desktop
PyUSB are undocumented and appear to be rather aribitrary.