Use PDM audio on esp32-c3-lyra board
The I2S data pin is hooked up directly to the amplifier on this board,
so while it does generate sound, it sounds terrible since it's not
actually handling the data signal.  The sample implementation in the
esp-adf uses the I2S hardware with PDM encoding to drive the amp.

https://github.com/espressif/esp-adf/blob/ad4ac707c5dffb450dbabfdcaa7443165ee4e852/examples/player/pipeline_spiffs_mp3/main/play_spiffs_mp3_example.c#L76

This change hardcodes the PDM implementation for this board.