switch usb buffer fix to static buffer
This modifies tannewt's DMA-safe buffer fix for usb.core.Device to
use a single global static buffer instead of potentially many heap
allocated buffers. Hopefully the shared buffer will avoid runtime
allocation errors.

CAUTION: The size of the buffer may not be big enough for some
transferst that people might want to make. If the requested length
of a transfer is too big, that length will be truncated to what can
fit in the buffer. The actual amount of transferred data will be
indicated by the length in the transfer result. The calling code
_should definitely_ check to see if the result matched what they
asked for.